(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=2834 Bug 2834: Tag queries need to be optimized
* Added indexes to the tags and flags columns in mail_item.

* Added a check for searches that would match every tag or flag,
so we don't generate a big IN clause.

* Fixed the case where a search for a tag or flag that doesn't
match anything incorrectly returns all the results.  Thanks
to Dan for finding this.

Separate change that got rolled into this one due to overlap
in TestTags: renamed the unread flag from "\Unseen" to "\Unread".
Added a unit test that makes sure that "is:unread" and
"tag:\Unread" return the same results.

Updated query.txt with info on searching by the flag name.

Copied from Perforce
 Change: 11427