(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=3483 bug: 3483
Store the server-side date (in milliseconds) when an item is changed.  Return this date as the "md" attribute on each modified item in a SyncResponse.  Also return "md" on the SyncResponse itself.

The DB schema version is now 16.  Please reset-the-world or run src/db/migrate20050822-TrackChangeDate.pl

Copied from Perforce
 Change: 11429