(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=1711 bug: 1711
Maintain the redo log sequence number in db.
On tomcat startup, this number is compared with that in redo.log. If they dffer, tomcat won't start up and offline restore needs to be run to reset the correct sequence number.

Need to run reset-the-world in development env.

Copied from Perforce
 Change: 11431