(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=3560 Bug: 3560
This bug was caused by new version of ical4j.  It apparently doesn't like seeing both Standard and Daylight properties having the same offset in a VTimeZone.  So we now check for that case and only add Standard property for a time zone that doesn't use daylight savings time.

Copied from Perforce
 Change: 11591