(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=3531 Bug 3531: sessions not being purged on dogfood
Ran some tests locally and came to the conclusion that
sessions are being allocated and removed as expected.
Added debugging information to SessionCache
which will help us track down whether the sessions
are hanging around due to a server bug or client requests.

Added a ValueCounter class, used for keeping track
of the number of occurrences of an object in a collection.

Renamed StringUtil to TestUtilCode (TestUtil is taken).
Easier to just put all the short util tests in one place
for now instead of creating a separate class for each
category.

Copied from Perforce
 Change: 11725