(split)PubICalServlet: public (ie no-authtoken-required) servlet which allows external users to reply to calendar Invitations by just hitting a URL.
Basically, when sending out the invitation, the client can specify URLs in the HTML like <A href="@@ACCEPT@@">Accept this invite</A> or <A href="@@ACCEPT@@">blahblah</A> ---> this allows non-ical-compliant email clients to respond to invitations.  The server patches those URLs with the appropriate URL for each action and each meeting Attendee.

The PublicICalServlet is a very simple stub servlet which handles the requests and sends the appropriate update email out to the Organizer and Attendee involved.

Copied from Perforce
 Change: 14957