(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=4392 bug: 4392
The server wasn't parsing and setting the 'fb' attribute when creating an appointment from the client.

Changed the server and the spec to expect "O" for Busy-Unavailable (OutOfOffice) -- because that's what the client was sending, and it seemed easier to just change the server and be done with it.

Copied from Perforce
 Change: 15489