(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=248 bug: 248
For some reason I kept thinking this was the same bug as 2937...but really, it was not.  In fact it was a totally separate issue with a trivial 1-line fix.  Maybe I'll pay more attention next time.

Anyway -- when returning the <SearchConv> results, the MORE flag is set if there are any more messages in the conversation -- it does NOT get set based on wether or not there are more hits (because SearchConv returns ALL messages in the conv, regardless of wether or not they actually match the Search string)

Copied from Perforce
 Change: 15936