(split)http://bugzilla.zimbra.com/show_bug.cgi?id=4244 http://bugzilla.zimbra.com/show_bug.cgi?id=4958 bug 4244 bug 4958
add global config/server attribute zimbraMailMode (can be
http, https, both or mixed - not clear to me that we
support mixed very well - hasn't been tested in a long
time, but it's there as an allowed setting).

also add zimbraMtaAuthHost and zimbraMtaAuthURL.  The
latter is set when you change the former with verification
that zimbraMtaAuthHost actually points at a real server
object.

QA alert - this change touches (but does not break
compatiblity or require new presence of new attrs for) code
that deals finding URLs for inter-store server communication.
So backup, cross mailbox search and delegated access on a
multi-node install should be tested.

Copied from Perforce
 Change: 18600