(split)http://bugzilla.zimbra.com/show_bug.cgi?id=4661 Bug: 4661
Never send an email to the appointment organizer during appointment create/modify/delete.

When getting appointment summary, use the appointment's default freebusy and partstat if requester is the organizer.  This prevents a meeting from showing up as tentative in organizer's calendar view when organizer is also listed as attendee.

Copied from Perforce
 Change: 18690