(split)wire up IMAPS and IMAP+STARTTLS over NIO.
timeout unauthenticated IMAP connections in 60 seconds and
authenticated ones in 30 minutes.

make ImapSession deal with both NIO and TcpServer variants of
connection handlers.  IMAP IDLE command still doesn't work in NIO case
- looking into it.

localconfig variable "nio_imap_enable" set to true will start NIO IMAP
and IMAPS.  Not ready for primetime yet (I've tested the NIO framework
standalone rather well, but the IMAP part of it hasn't been beat on),
so don't enable it on dogfood and start looking for the country code
for India.  Also note that there is no "d" in "nio_imap_enable" - that
one bit me for a while.  You can modify localconfig-dev.xml in your
dev build and r-t-w to turn this on if you want to play with it.

Copied from Perforce
 Change: 19306