(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5132 Bug 5132: We shouldn't use GROUP_CONCAT in our database code
Updated markDeletionTargets() to not use GROUP_CONCAT.

Old logic: When
a folder deletion or purge occurred, we would prepend each
conversation's metadata with the list of its deleted messages.
The next time the conversation was instantiated we would use
that id list to recalculate the sender list.

New logic: we simply null out the metadata of any affected conversation.
The next time the conversation is instantiated, it recalculates
its metadata from scratch.  The new logic requires a little more work.
Dan and I agreed that this is a pretty uncommon operation, so there
shouldn't be any performance fallout.

Copied from Perforce
 Change: 19771