(split)http://bugzilla.zimbra.com/show_bug.cgi?id=3824 Bug: 3824
Added versioning to redolog file header.  Version is now set in the file header and in each redo record.  Incremented version number to 1.0.  (Previous redolog version number was 0.1.)

Added list of mailbox IDs to StoreIncomingBlobs redo record, for version 1.0.  This allows restore code to replay the op only if the blob is relevant for the mailboxes being restored.

Copied from Perforce
 Change: 20321