(split)http://bugzilla.zimbra.com/show_bug.cgi?id=4976 Bug: 4976
Correct handling of UNTIL in recurrence rules.

UNTIL value is inclusive, and can be specified as time-less date (e.g. "20060125") or UTC date-time, one that ends with "Z".  The UTC time case is simple; if meeting instance's start time is less than or equal to UNTIL, it's a good instance.  To get the date-only case to work correctly, it's necessary to extend the actual UNTIL time to 23:59:59 of the specified date in local time zone, meaning the effective UNTIL date-time is original UNTIL + 1 day - 1 second.

Copied from Perforce
 Change: 20554