(split)Notification Reliability, checkpoint of instant-notification-port.
The server now sequences <notify> blocks and the client communicates a "last notification received" number up to the server when it connects: this allows for the retransmission of notifications that might be missed.

Checkpoint of the "instant notification port" -- the server listens on an alternate port and will accept HTTP requests to that port (with a notification seq no in them) and will let the connection "hang" until there is new data to send: effectively allowing for instant notification of server changes (no need to wait for a polling interval).

Copied from Perforce
 Change: 20998