(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5885 Bug: 5885
When expanding recurrences with UNTIL specified, look for bogus end time of "T000000Z" set by Outlook.  When it's found, or if the UNTIL is date-only, extend the cutoff time to the end of the day (23:59:59) in the time zone used in DTSTART (meeting start time).

This ensures the meeting instance on the day, which has the start time later than midnight (00:00:00) but at or before 23:59:59, is included in the recurrence expansion.

Copied from Perforce
 Change: 21165