(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6023 bug: 6023
Boris' change made things much simpler by nulling out the METADATA on a conversation when any of its messages were deleted as part of a folder delete.  Unfortunately, we continued to assume that mData.metadata would be non-null, and the Conversation's decodeMetadata threw an NPE every time it hit one of those.  Fix this by, uh, checking for null and recalculating the metadata when appropriate.

Copied from Perforce
 Change: 21332