(split)http://bugzilla.zimbra.com/show_bug.cgi?id=1246 bug: 1246
Handle inlined UUENCODEd attachments.  We leave the message unaltered on disk, but when fetching (and when indexing) the message, we remove the uuencoded data from the message's main text/plain part and present the message as a multipart/mixed with the remainder of the data in the text/plain part and the decoded attachment in an application/octet-stream part.  Multiple uuencoded files in a single message are supported.  uudecoding in attached message/rfc822 parts is supported.

Open issue: application/octet-stream attachments are not indexed in the OSS version.  (Are they indexed in the Network version?)  We could use .mime-types to guess at the MimeHandler based on file extensions.

Copied from Perforce
 Change: 21445