(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6020 bug: 6020
Add vCard support for creating contacts.  The vCard can either be uploaded and specified as an attachment, inlined in the <CreateContactRequest>, or referenced as an attachment of an existing message in the store -- see soap.txt for the syntax.  vCard 3.0 and 2.1 formats are supported, though the syntax checking is *very* loose at the moment.

Currently, when specifying message-id/part you must specify a numeric message id.  Need to allow the caller to qualify this id so that they can refer to messages in another user's store.

Parag, you should now be able to add an "add to address book" link next to vCard attachments.

Copied from Perforce
 Change: 21499