http://bugzilla.zimbra.com/show_bug.cgi?id=3657 bug: 3657 (Integration to CRAY) - Modify zmtrainsa to do auto-extraction and training from systemwide spam/!spam accounts. Invoke with no arguments to use this. - Add cron entry on mta to invoke this - Add cron enty on ldap server to invoke with --cleanup to empty out spam/!spam accounts.
Initially, I was going to train centrally, and then distribute to all mtas, but that would 
require including the as software (SA and dspam) on the training server (ldap was the logical
choice) which is not worth the trouble.  So, I'm training per host, from the central db, which 
should work ok.

Note - previous zmtrainsa functions still work fine, and both continue to train dspam (which is 
disabled in amavis by default)

Copied from Perforce
 Change: 22467