(split)http://bugzilla.zimbra.com/show_bug.cgi?id=1363 bug: 1363
Allow messages, etc. to be added to a folder via a POST to the REST interface.  Specify the folder to upload to in the REST URL and optionally the type (vcf, ics, csv) with a "fmt=" parameter.  Put the body of the message/vcard/icalendar/csv in the POST body and away you go!

Issues:
  - we blindly read the POST body into a byte[] in memory
  - can't yet specify tags, flags, date on added items
  - defaulting NativeFormatter and SyncFormatter to importing messages

Copied from Perforce
 Change: 22679