(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5975 Bug 5975: Daimler server stat monitoring
Changed zimbrastats.csv logic to use log4j.  The file was mysteriously
being truncated on dogfood.  Hopefully the log4j code will be robust
enough to handle whatever weirdness was happening.

Added a ZimbraStatsAppender class, which subclasses log4j's
DailyRollingFileAppender and adds a header to the top of every
new csv file.  Currently this appender is tied to zimbrastats.csv.
We can generalize it later if we want to use it for other files.

Updated collection classes in ZimbraPerf to be type-safe.

Copied from Perforce
 Change: 22822