(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6653 bug: 6653
Each Formatter can optionally have a set of Content-Types for which it is the default.  When a POST is received without an explicit "fmt=..." query parameter, we check the POST's Content-Type header and try to figure out the correct formatter from that.  If no matching formatter is found for the Content-Type, we fall back to the default for the target folder.

Copied from Perforce
 Change: 23109