(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6243 bug: 6243
MimeVisitor in now an abstract class that contains Mime.accept().  (Yeah, this will probably be undone at some point in the near future.)  We have two implementations of MimeVisitor: UUEncodeConverter and TnefConverter, both of which should be fully functional.  All converters inspect all message parts, which may be a little inefficient.

MimeVisitors fall into two classes: converters, which don't affect the disk representation of a message but are run on a message copy at fetch and index time, and mutators, which change a message before it's sent or saved to disk.  You register converters with MimeVisitor.registerConverter() and mutators with registerMutator().

Added a MimeMessage subclass that doesn't change the Message-ID header when saveChanges() is called on it.  Used this new subclass in a whole bunch of places.

Copied from Perforce
 Change: 23467