(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6243 bug: 6243
Make sure to run mutators before analyzing or getting raw data for messages passed in as MimeMessages.  This is performed at parse time, not at constructor time, which is not how it works when the message is passed to the constructor as a byte[] array, but it should be good enough.

Don't re-run non-applicable converters while parsing a message for indexing.

Copied from Perforce
 Change: 23699