(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6807 Bug: 6807
Decouple moving of a mailbox to a new server and deletion of old mailbox on old server.

-sw (switchover) option has been removed from zmmailboxmove command.  It will always switch over to new server after successful move by updating zimbraMailHost attribute of the account in LDAP.

To move a mailbox, invoke zmmailboxmove with "-a <email>" and "-t <target server hostname>" options.  This *copies* the mailbox from current server to new server.  The old mailbox is left intact, but all users, even those whose session was open at the time of mailbox move, will get wrong-host error on the next operation.  Thus the user is forced to switch to the new server immediately, making it impossible to make further updates to the soft-deleted mailbox on the old server.

At this time sysadmin or user has a chance to inspect the new mailbox for data integrity.  If satisfied, admin can issue another zmmailboxmove command, this time with "-po" option (Purge Old) along with "-a <email>" to permanently delete the old mailbox from old server.  The location of the old mailbox is tracked in LDAP, in zimbraMailboxLocationBeforeMove attribute.  It has the format of "<server zimbraId>:<mailbox ID on old server>".  This attribute is set on mailbox move to new server (-t option), and cleared on purge of old mailbox.  Leaving the old mailbox around will not affect the operation of new mailbox.  It just wastes space on the old server.

Here's the new usage output of zmmailboxmove command:

usage: zmmailboxmove -a <email address> [ -t <target server> OR -po ]
 -a,--account <arg>        Email address of account to move.
 -h,--help                 Displays this help message.
 -ow,--overwriteMailbox    Overwrite target mailbox if it exists
 -po,--purgeOld            Purge old mailbox on previous server (for
                           successfully moved mailboxes)
 -s,--server <arg>         Mail server hostname. Default is localhost.
 -t,--targetServer <arg>   Server where mailbox is moved to

-t and -po options are mutually exclusive.
Use -t option to move a mailbox from the current server to the new server
specified by the option.  If -ow is specified, any existing mailbox for
the account on the target server is overwritten.
After successfully moving the mailbox, use -po option to purge the old
mailbox on the original server.

<end of usage>

Usage output doesn't mention -a is always required and one and only one of -t or -po is required.  But this becomes evident from the error message printed when invoked with wrong set of options.

-ow option can be used only when -t is used.  For whatever reason, you might find there's already a mailbox on the new server for the user specified in -a option.  If -ow isn't used, the move will fail with meaningful error message.  You can then specify -ow to wipe out that mailbox and force the move of the mailbox you want to move.

-s option can be used with both -t and -po.  It specifies to which server the zmmailboxmove CLI process sends the SOAP request.  By default the requests are sent to the admin servlet on localhost, but depending on which machine is old server and which is new server for the mailbox being moved, you can use the -s option to have the requests sent to the right servers.  And the requests must be sent to the right servers.  If not, error will be printed on command line.

Example: Suppose you want to move mailbox for user1@zimbra.com from current server foo to new server bar.  You are invoking zmmailboxmove on foo.

Step 1: zmmailboxmove -a user1@zimbra.com -t bar
  (mailbox is copied from foo to bar; all new requests for mailbox go to bar)
Step 2: Test the new mailbox on bar.
Step 3: zmmailboxmove -a user1@zimbra.com -po
  (old mailbox on foo is purged)

Another example: You are still moving user1@zimbra.com from foo to bar, but you want to invoke zmmailboxmove on bar instead.

Step 1: zmmailboxmove -a user1@zimbra.com -t bar -s foo
  (you have to send the move command to the old server; it drives the copy operation by coordinating with the new server; zmmailboxmove never talks to new server directly)
Step 2: Test new mailbox on bar.
Step 3: zmmailboxmove -a user1@zimbra.com -po -s foo

Copied from Perforce
 Change: 23899