(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7079 Bug: 7079
When SendInviteReply request specifies id of an invite email for an exception instance to a recurring appointment, pay attention to the Invite object's recur ID and do the accept/decline only for the instance in question.  Don't accept/decline the entire series like the old code was doing.

Copied from Perforce
 Change: 24112