(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6582 bug: 6582
When we're in Windows Mobile 5 IMAP hack mode and we see a multipart/alternative part, replace the multipart by promoting its primary subpart.  There should be no multipart/alternatives in the resulting BODYSTRUCTUREs.

Unfortunately, we run up against broken JavaMail caching, so in order to get this to work we need to serialize the resulting message to a buffer and then read that buffer back into a new MimeMessage.  This is absolutely terrible, but necessary unless and until I can figure out how to work around JavaMail.

Copied from Perforce
 Change: 24260