(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6606 bug: 6606
Fix a problem with the DBQueryOperation combining logic that was causing them to incorrectly rewrite to NullQueryOperations in some cases -- specifically when a ITEM: (or CONV:) tried to combine with something else.  This would cause a query like "in:inbox conv:123" to return no hits even if it should have...and since none of the messages in the conversation were a match, the fetch would fail.

Copied from Perforce
 Change: 24529