(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6651 bug: 6651
Untagged NOOPs for COPY are now done based on time (15 secs of inactivity) rather than on number of messages copied, and henceforth this will be the standard.  Move expungeMessages() to the handler so we can more easily issue untagged NOOPs to keep Thunderbird from timing out.  Enabled untagged NOOPs for silent STORE and issue untagged FETCHes when they occur for non-silent STORE.  Reorg the STORE code to hold the mailbox lock for as little time as possible; expunging is still done with a lock for the entire time to keep new message notifications for occurring.

Tweak the ImapFolder code to implement Iterable<ImapMessage> (in the normal sequence order).  Move the "notifications disabled" check into ImapFolder.dirtyMessage(), where it belongs.

Copied from Perforce
 Change: 25132