(split)http://bugzilla.zimbra.com/show_bug.cgi?id=4210 bug: 4210, 6019, 6954
Add an IMAP_ID column to the user mailbox database; this defaults to the item ID for most leaf-node types.  The ZIMBRA.MAILBOX table now has a TRACKING_IMAP boolean which is set the first time a user logs in via IMAP.  The TRACKING_SYNC column in ZIMBRA.MAILBOX is now an INTEGER; we will be using this in the future to track trimming the TOMBSTONE table.  Database version is now 23; please reset-the-world or run the script ZimbraServer/src/db/migrate20060515-AddImapId.pl to bring your build into the new world.

When loading non-search folders in IMAP, we now directly fetch only the relevant fixed-width columns from the database (ID, IMAP_ID, TYPE, FLAGS, TAGS, UNREAD) and don't instantiate any MailItems in the Mailbox.  Items that have an IMAP_ID <=0 have a new IMAP ID assigned at folder load time; this state occurs automatically when an item is moved via a non-IMAP interface when TRACKING_IMAP is true for the account.  Certain message attributes that are seldom fetched without fetching the message body (date, size) are no longer precached when the IMAP folder is loaded but are instead retrieved by fetching the item on demand from the Mailbox.

Expose contacts via IMAP as messages with a text/x-vcard Content-Type with UTF-8 content.  The Contacts folder is readable but not writable.

Move IMAP ID up to MailItem.  Also collect blob-writing code in MailItem.setContent(), so we're no longer duplicating the code in Message and Appointment.  Note that Document does not yet share this blob-wrangling code.

vCards are now \r\n-delimited.  Please let me know if this breaks anything...

Copied from Perforce
 Change: 25480