(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7691 bug: 7691
When auth is set to internal, and password isn't in SSHA format, then attempt to bind to the LDAP server as the user/password instead of failing. This will make things easier for customers that have imported passwords that are in other formats, like MD5/CRYPT/etc.

Copied from Perforce
 Change: 25643