(split)http://bugzilla.zimbra.com/show_bug.cgi?id=2085 bug: 2085
Allow specifying a new set of contact fields as part of ContactActionRequest with op="update".  Just inline the <a> fields inside the <action> element and we'll replace the existing contact's fields with the supplied ones.

It'd make sense to get rid of ModifyContact, but that'll wait for later.

Copied from Perforce
 Change: 25663