(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7459 http://bugzilla.zimbra.com/show_bug.cgi?id=6536 Bug: 7459, 6536
Implement a hack to keep Microsoft Entourage happy.  Entourage has a bug that causes iCalendar parsing to fail when the TZID parameter in DTSTART/DTEND is not surrounded by double quotes, even when quoting is not necessary according to iCalendar spec.  To get around this, we now always quote TZID when it's used as a parameter of a property.  It is not quoted when TZID is a property rather than a parameter.  ("property", "parameter", etc. are iCalendar terms)

Always set DESCRIPTION property in VEVENTs.  This is another Entourage quirk.  It won't display the text body provided as multipart/alternative parts.  Only the DESCRIPTION of iCalendar is rendered.

Implemented line folding at 76-chars for iCalendar MIME parts generated.

Use CRLF as line break in iCalendar part, rather than bare LF.

Copied from Perforce
 Change: 26239