(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7309 Bug: 7309
When modifying appointment series to a new start time, update all exception instances to reflect the new time in the RECURRENCE-ID.  This ensures the exceptions continue to be in place after series modification.  End result is that only unmodified instances are moved to the new time.  Exception instances will not be changed in any user-visible way (time, subject, notes, etc.) by the series edit.

Copied from Perforce
 Change: 27082