(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5343 Bug: 5343
In GetApptSummaries response, for each instance start time return the time zone offset at that time in the TZ of the appointment.  This information enables the client to render an all-day appointment on the correct day when the appointment's time zone is different from the client's local time zone.

Copied from Perforce
 Change: 27092