(split)http://bugzilla.zimbra.com/show_bug.cgi?id=8059 bug: 8059
Reimplement IMAP COPY as a move to the target folder followed by a copy back to the original folder, with the copied message not part of the original message's conversation (if any).  The moved message is assigned a new IMAP UID; the copy is assigned the original message's IMAP UID and should seamlessly replace the original message in the SELECTed IMAP folder.  (That's what "ghosting" is for -- noting that a move/copy is about to occur, so the session's notification handler knows to treat the move/copy notifications specially.)

"added" and "expunged" are now implemented as session flags rather as distinct booleans.  They're also no longer set directly from client code.  Yay for getters/setters.

If an IMAP COPY fails halfway through, we're not rolling it back.  That's not correct, but that's what we're doing for now.

Differentiated "mutable" and "deletable" on MailItems.  Non-draft messages are immutable but deletable.  System folders are immutable and undeletable.  Pretty much everything else is mutable and deletable.  When a mutable item is copied (either really copied or the IMAP move/copy combo), we index the copy separately.  [Should we get tricky and do reindex on write?]

Copied from Perforce
 Change: 27424