(split)http://bugzilla.zimbra.com/show_bug.cgi?id=8121 bug: 8121 - Log in new account via web creates infinite loop
An exception thrown during mailbox creation is not being logged.
I suspect that something is going wrong in the rollback logic,
which causes the original exception to be lost.  Added catch
blocks with extra logging to Mailbox.createMailbox() so that
we have more information the next time this happens.

Copied from Perforce
 Change: 27944