(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7046 http://bugzilla.zimbra.com/show_bug.cgi?id=8569 bug: 7046, 8569
Modify Mailbox.search() so that the caller can request the SoapProtocol they want the response in (relevant for Proxied search requests).  This means we no longer have to convert ProxiedHits from XML->JSON.

Mailbox.search() no longer allows an OperationContext of NULL to be passed-in.  Modified all callsites to create one of they were previously passing in NULL.

If a proxied search request generates an error, just pass the SoapFaultException straight up, don't wrap it in a ServiceException -- we can do this now that SoapFaultException has been changed to extend ServiceException.

Copied from Perforce
 Change: 28082