(split)http://bugzilla.zimbra.com/show_bug.cgi?id=8491 bug: 8491
Enforce mailbox permissions in search requests.  This is enforced by calling Mailbox.getVisibleFolders(octxt) and then adding the search constraints "and (in:folder1 or in:folder2 or in:folder3)" to the local part of the search we're running (permissions for remote parts are checked when they're executed on the remote server).

Also fixed a bug in the query optimizer -- searches like "in:inbox and in:otherfolder" were returning results b/c the optimizer was failing to set the no-result flag.  Oops.

Copied from Perforce
 Change: 28380