(split)http://bugzilla.zimbra.com/show_bug.cgi?id=9560 bug: 9560
Rewrite the static Mailbox.beginMaintenance() to not hold the Mailbox cache lock while calling synchronized Mailbox methods.  We're faulting in a Mailbox before doing the op (if it's been lazily created in the DB), which simplifies the code a bunch.

We also use the SoapSession's already-cached mMailbox instead of fetching it from the Mailbox cache during notification.  So we've hit this issue from both sides, and it should be resolved.

Copied from Perforce
 Change: 29684