http://bugzilla.zimbra.com/show_bug.cgi?id=9805 bug: 9805
integrated from edison.
the call to ChangeManager::MoveMessages() was made too early....before the inner call to CopyMessages(). the problem was that MoveMessages() discovers the item's parent folder to find out where to put the change record for the move.  since this needs to be in the destination folder, we need to make the call to MoveMessages() after the call to CopyMessages().

Copied from Perforce
 Change: 29808