(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5003 http://bugzilla.zimbra.com/show_bug.cgi?id=2809 http://bugzilla.zimbra.com/show_bug.cgi?id=7906 bug: 5003, 2809, 7906
Persist unread/item counts for folders and unread counts for tags in the database.  Also persist total size/contact count for a mailbox.  This is done on a per-item basis at the end of every write transaction where these counts change.  Startup read times should now be significantly faster, at the cost of more overhead for most write transactions.  Quota usage reports should now be accurate.

Folder "size" is now the number of non-subfolder items in the folder.  (It now works the same way that conversation size works.)  As a result, MailItem.getSize() can now return an int and every MailItem shrinks by 8 bytes.

To make this possible, MAIL_ITEM.UNREAD is now an INTEGER and there's a new CONTACT_COUNT column in ZIMBRA.MAILBOX.  As a result the database version is bumped to 27 for mainline; please reset-the-world or run ZimbraServer/src/db/migrate20060810-PersistFolderCounts.pl to get your DB schema in sync.

We now also support IMAP UIDNEXT, both on SELECT/EXAMINE and STATUS.  This data is maintained per-folder and updated whenever the "size" (item count) of a folder increases in the same write that updates the size in the folder DB row.  It is tracked regardless of whether the user has used IMAP or not.

Copied from Perforce
 Change: 29948