(split)http://bugzilla.zimbra.com/show_bug.cgi?id=9931 bug: 9931
Explicitly disallow folder/tag/conversation matches when fetching items with UNREAD > 1 for marking read.  These items used to all have UNREAD = NULL, but with the recent changes for bug 5003, folders and tags non have non-NULL UNREAD.

Treat the UNREAD column as an INTEGER instead of a BOOLEAN across the board, 'cause that's what it is now.  MySQL does the right thing (i.e. treats it as 0 or 1), but don't rely on that.

Copied from Perforce
 Change: 30026