(split)http://bugzilla.zimbra.com/show_bug.cgi?id=9889 bug: 9889
Support text/enriched body parts by converting to HTML on the fly before text extraction and when returning the message to the client.  All of RFC 1896 should be covered except <ParaIndent>.

Both text/plain and text/enriched parts are viewed as alternates when looking for an HTML "body part", and both text/html and text/enriched parts are viewed as alternates when looking for a plain text "body part".  We now prefer the latter of two possible alternates in the MIME structure.

zimbra_mimehandlers.ldif is updated by this change both for OSS and Network.  If any install/upgrade changes must be made to accommodate this, this change has *not* done so.

Copied from Perforce
 Change: 30099