(split)http://bugzilla.zimbra.com/show_bug.cgi?id=2414 bug: 2414
Added a field to CreateMessage redo op to track the ID of the message which, together with the message being created, forms a new conversation.  In the code this is called "first message of conversation".  This information is used during restore to update the conversation ID of that message.

Bumped up redolog version to 1.5.

Copied from Perforce
 Change: 30387