(split)http://bugzilla.zimbra.com/show_bug.cgi?id=10652 bug: 10652
Moved iCalendar REPLY processing out of Mailbox.addMessageInternal().  This allows REPLY processing to occur with the mailbox not synchronized and no transaction started, making it possible to synchronize or start trasaction on another mailbox, that of the organizer on whose behalf the current mailbox is receiving the calendar reply.

Copied from Perforce
 Change: 32142