(split)http://bugzilla.zimbra.com/show_bug.cgi?id=9622 bug: 9622
Added (and documented) the new SOAP command GetMsgMetadata, which returns the size/date/etc. for a set of messages but does not return the subject/sender/fragment.

Cleaned up GetMsgOperation and made it optionally take a list of message IDs.  Removed the call to schedule() from the constructor.

Copied from Perforce
 Change: 32304