(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11168 bug: 11168
Hide IMAP \Deleted messages from POP3 sessions.  The RFC doesn't say to do this (and it doesn't say *not* to do this), but it hides an issue customers were seeing due to the way IMAP COPY is currently implemented.  The problem is that IMAP COPY is done using a move and a copy back to the original folder, which makes the ID (and hence the UIDL) on the "copied" message appear to change.

Copied from Perforce
 Change: 33018