(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5787 bug: 5787
During yearly rule expansion, don't end the while loop at the beginning by comparing the iteration's base date and range end date.  The check may end the loop too early if DTSTART doesn't satisfy the recurrence rule and DTSTART's day-of-year is later than the day-of-year of the UNTIL date, in a yearly recurrence.

Instead, break out of the loop at the end of loop body if all possible instances examined in the iteration are later than the range end date, which is equal to the UNTIL date in the final year.  This relies on the fact that the instances examined during the entire loop are in strictly increasing time order.

Copied from Perforce
 Change: 34065