(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11636 bug: 11636
StoreIncomingBlob operation's timestamp was being incorrectly recorded in seconds, when it should have been recorded in milliseconds.  This causes the operation to look like it happened in distant past and is not replayed.  Then a subsequent CreateMessage replay tries to link to a blob file that would've been saved by StoreIncomingBlob, which never happened, and hence the missing blob error.

This fix only addresses the problem for future redologs and incremental backups.

Copied from Perforce
 Change: 34591