(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11733 bug: 11733
Truncate the "write-only" SENDER column to 128 characters when writing to the database.  Note that this value (message's sender, contact's fileAs) is used only for sort purposes and is never read out of the database.

This reintroduces some index/database sort issues, as the index is not truncating these values and thus may sort differently from the database in the very very very rare instance when the SENDER value is truncated.  We probably need to install the same checks in the indexing code; I'm leery of putting the truncation in MailItem.

Copied from Perforce
 Change: 34602