(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11718 bug 11718
add zimbraAllowFromAddress (multi-value email on account) and
zimbraAllowAnyFromAddress (boolean on account/cos, inherited).  If
the allow any boolean is set, the server allows any From addess in
SendMsg.  If the boolean is false, then you are allowed to use one of
your addresses or an address set in zimbraAllowFromAddress.

Note that the neither attribute can be set by the end user and have
to be set by the admin.

The UI should change to allow a text entry field if the "allow any"
boolean is true, and to show a drop down list (populate with account
address + aliases + allowed addresses) if the boolean is false.

Bug should also be moved to admin UI after the end user UI to allow
admin editing of these variables.

Copied from Perforce
 Change: 34969