(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6852 bug: 6852
WM5 seems to always want a "CHARSET" parameter in the BODYSTRUCTURE representation of a text/plain part, so force it to "us-ascii" if it's not already set.  This is contrary to RFC 3501 section 7.4.2, but it seems sufficiently innocuous that I'm not making it dependent on being in WM5 hack mode.

Copied from Perforce
 Change: 35306