(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11937 bug: 11937
Queries can now be executed in DB_FIRST order regardless of the size of the DB result set.  This should fix the "is:unread" searches.

Added a temporary heuristic to the query engine to make it choose a DB_FIRST order if the Lucene part of the query has only one term and that term matches more than 5000 items...

Copied from Perforce
 Change: 35322