(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12210 bug: 12210
Add SENT-BY parameter to ORGANIZER/ATTENDEE properties of an iCalendar part in the MIME message of <SendMsgRequest> when a user is sending on behalf of another user, indicated by needCalendarSentByFixup=1 attribute.

SENT-BY is added in any ORGANIZER and/or ATTENDEE and set to the email address in Sender message header if the property value is the same as From address.

Copied from Perforce
 Change: 35473