(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12242 bug: 12242
Support invites with no organizer, whether or not attendees are present.  Allow changing the organizer between being unset and being set to self, because this is possible when the meeting creator user adds/removes attendees.  We continue to block changing organizer to another user because there is no good way to tell the request is not malicious.

Copied from Perforce
 Change: 35570