(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7964 bug: 7964
Add the concept of an optional "send UID" to the SendMsg request.  If the caller specifies a "send UID" on the <SendMsgRequest> element, we'll track the progress of the message send.  If the caller doesn't receive the SendMsgResponse (or the client "aborts" the send), they can re-send the same SendMsgRequest with the same "send UID" and the server will try to use the result from the previous SendMsg call rather than sending a second copy.

Clients should do their best to make their "send UID" unique for the user.  A timestamp should be sufficient.

If the message is sent, "aborted", **edited**, and re-sent, the client should use a different "send UID" since the message has changed.

We track the status for the last 5 SendMsg calls with "send UID" specified.  If a second SendMsg comes in while the first is still processing, we will wait for up to 4 seconds for the original SendMsg to complete.  If it's still incomplete after 4 seconds, the server will throw mail.TRY_AGAIN.

Copied from Perforce
 Change: 35725