(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12313 bug: 12313
Support RFC 4731 (IMAP4 Extension to SEARCH Command for Controlling What Kind of Information Is Returned).

SEARCH and UID SEARCH can be passed "RETURN (...)" as the leading parameter; if so, the resulting untagged responses are "ESEARCH" instead of "SEARCH" and contain only the requested information.  Supported RETURN options are MAX, MIN, ALL, and COUNT.  If RETURN is included but no options are specified, we treat it as if ALL was requested.  If no reults match the SEARCH query, MAX, MIN, and ALL are omitted from the results.

Copied from Perforce
 Change: 35738