(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12419 bug: 12419
Update supported SEARCHRES extension to -04.  We're exposing "SEARCHRES" as the extension capability instead of the "X-DRAFT-04-SEARCHRES" capability specified by the draft in the hopes that the syntax and semantics remain unchanged.

Main changes from -03 to -04 are that the saved search result defaults to the initial contents of the folder at SELECT time and that (MIN SAVE) and (MIN MAX SAVE) and (MAX SAVE) without COUNT or ALL only save the specified MIN/MAX values.

Copied from Perforce
 Change: 36704