http://bugzilla.zimbra.com/show_bug.cgi?id=11933 bug: 11933
Eliminate the separate lock on the index: just use the Mailbox lock instead.

For now, the localconfig key 'debug_mailboxindex_use_new_locking' (defaults to TRUE) controls if we use the new style (always lock the Mailbox) or the old style (separate locks for Mailbox and Index) -- this is intended for testing so that we can do some before and after perf tests to make sure this doesn't hurt performance significantly.

Copied from Perforce
 Change: 36837