(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12173 bug: 12173
The caller of a SOAP method that creates mail (e.g. Save Draft, SendMsg, any of the group calendaring methods) can specify the identity that they're sending the message from by adding an optional

    idnt="{identity-id}"

attribute to the request's <m> element.  At present, this controls the save-to-sent behavior of the message send; in the future, we'll extend this to allow the tagging/flagging/etc. of sent messages.

When a draft is saved with an identity specified, subsequent GetMsgResponses for that message will include the specified "idnt" attribute on the returned <m> element.  Note that at present the identity cannot be changed one it has been initially set.  Also, the identity ID is not yet stored in the SaveDraft redolog operation.

Copied from Perforce
 Change: 36888