(split)http://bugzilla.zimbra.com/show_bug.cgi?id=11779 bug: 11779
Implemented POP3 leave on server.  When a POP3 data source
specifies leaveOnServer="true", we now check every
message's UID against the UID's stored in the pop3_message
table for that data source.  If we've already encountered
the UID, we don't retrieve the message.  If it's a new
message, we add it to the mailbox and store the UID.

If the host, username or leave-on-server setting of a
POP3 data soruce is modified, we assume that this is
effectively a new data source and wipe out the UID list
for that host/user.

* Added leaveOnServer attr to data source SOAP messages
* Added pop3_messages table
* Added DbPop3Messages class for modifying the table
* Updated ZMailbox API
* Added unit tests for testing POP3 UID persistence

*** NOTE: MIGRATION IMPACT ***

* Added migrate20061204-CreatePop3MessageTable.pl script

Additional stuff:

* Moved test running code from ZimbraSuite to TestUtil to allow
tests without server dependencies to run from the command line.
* Cleaned up compiler warnings in util tests.

Copied from Perforce
 Change: 37114