(split)http://bugzilla.zimbra.com/show_bug.cgi?id=12780 bug: 12780
Force all Mailboxes to reindex their contacts on logon if the Mailbox-version is < 1.2 (they all are).  For the few installs that have been running off of head (ie dogfood), this will cause a second reindexing of contacts now that Dan's script has been run -- for all other installs, this reindexing will be consolidated into a single job with the existing reindexing (mailbox v 1.0 -> 1.1).

Copied from Perforce
 Change: 37742