(split)http://bugzilla.zimbra.com/show_bug.cgi?id=6582 http://bugzilla.zimbra.com/show_bug.cgi?id=12998 bug: 6582, 12998
Support the extra data returned for BODYSTRUCTURE.  It's "optional", but optional in that you can skip it if there's no data to return.  At some point we may want to remove all the trailing NILs (often 4 of them), but at least we should now be correct.

The hope is that this resolves the WM5 message display issues by letting the client know the boundary string for multiparts.

Copied from Perforce
 Change: 38070