(split)bug: 9622
Add a new GetItem operation that allows a SOAP client to fetch an item of unknown type either (a) by id, (b) by absolute path, or (c) by name within a parent folder.  (We may want to extend this to other Get* SOAP operations.)

The returned item is serialized normally, so it's an <m> when the requested item is a message, <folder> when it's a folder, etc.

Copied from Perforce
 Change: 39052