(split)bug: 12365
Support the -03 version of the WITHIN draft, which requires that the argument to OLDER and YOUNGER is a number of seconds (was hours in earlier drafts).  To perform this search internally, we use the variant of the "berfore:" an "after:" queries that take a Java time (milliseconds since the epoch).

Copied from Perforce
 Change: 39150