(split)bug: 12531 - If "Use SSL..."is not checked to pop gmail, test-connection hangs application
JavaMail was ignoring the socket timeout because I passed it as an
integer to Properties.put() instead of as a String to
Properties.setProperty().  Oy.  Updated the code to call the correct
API and confirmed that the timeout is now working.  Bumped the timeout
up to 20 seconds so we don't hit problems with slow networks.

Copied from Perforce
 Change: 39342