(split)http://bugzilla.zimbra.com/show_bug.cgi?id=7074 bug: 7074
Populate WF's list of local domains by calling Provisioning.getAllDomains() at startup.  Validate incoming client connections against the list of domains and reject the client with 'host-unknown' if the client is from a domain we don't know about.  There is currently no listener on the list of domains -- this means that you need to restart the server when you add/remove domains for now.

Copied from Perforce
 Change: 40873