(split)http://bugzilla.zimbra.com/show_bug.cgi?id=5554 bug: 5554
Default status of tasks should be NEEDS-ACTION, not CONFIRMED which is the default for appointments.

For tasks, don't use Outlook compatibility hack for date-only values.  The hack is still in place for appointments, to convert date-only values to date/time with time="T000000" in the local timezone.  This hack is required to keep Outlook happy, but isn't necessary for tasks because Outlook doesn't support VTODO in .ics objects.

For appointments, we force all-day DTEND to be greater than DTSTART.  Don't do this for DUE field (equivalent of appointment DTEND) in tasks.  Allow DUE to be same as DTSTART or even earlier.

Copied from Perforce
 Change: 42173